docs: sd-divider docs [skip chromatic] (#1255)
